+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # BUG FIXES
 
 - 客户端管理 错误的status disabled
+- modal/drawer升级后zIndex(2000)会遮挡Tinymce的下拉框zIndex(1300)
 
 # 1.2.0